refactor(*) remove a level of indirection
Methods to a struct are aplicable to either the struct or a pointer to
it.  Channels are already reference
types, so no significant gain by adding another indirection.

Methods to pointers, OTOH, apply to a struct only if it is
"referencible", and should be used only if the method
modifies the given value.  (using a channel isn't considered
modification)